+++ modulo4/labecommerce/src/data/2_8.a-selectAllUsers.ts
@@ -7,4 +7,4 @@ export default async (): Promise<[]> => {
   `)
 
   return users
-}; 
+};
