--- modulo4/labecommerce/src/endpoints/2-getAllUsers.ts
@@ -1,18 +1,0 @@
-import { Request, Response } from "express";
-import selectAllUsers from "./../data/2-selectAllUsers";
-
-export async function getAllUsers(req: Request, res: Response): Promise<void> {
-  try {
-    
-    const users = await selectAllUsers();
-
-    if (!users.length) {
-      res.statusCode = 400;
-      throw new Error('Existem dados faltantes!');
-    }
-
-    res.status(200).send(users)
-  } catch (error) {
-    res.status(500).send("Internal server error")
-  }
-}
