+++ modulo4/labecommerce/src/index.ts
@@ -1,6 +1,6 @@
 import { app } from "./app";
 import { createUser } from "./endpoints/1-createUser";
-import { getAllUsers } from "./endpoints/2-getAllUsers";
+import { getAllUsers } from "./endpoints/2_8-getAllUsers";
 import { createProduct } from "./endpoints/3-createProduct";
 import { getAllProducts } from "./endpoints/4_7-getAllProducts";
 import { createPurchase } from "./endpoints/5-createPurchase";
